refactor(pt_expt): hoist graph with-comm exportable into make_model
LinearEnergyModel compositions need it once the with-comm gate opens for
them (issue #5906 Task 4); one owner, next to the non-comm twin.
_translate_energy_keys moves along (make_model cannot import from
ener_model without a cycle); importers re-pointed.
